refactor: 알림 이벤트를 알림페이지로 이동
수정사항
- ApplyButtom.js에서 진행하던 알림 이벤트를 알림페이지로 이동
